feat: user methods implementations

+++ src/context/users/deleteUser/deleteUser.controller.ts
@@ -1,16 +1,18 @@
 import {
+  Body,
   Controller,
   Delete,
   HttpCode,
+  HttpException,
   HttpStatus,
-  Query,
+  Response,
 } from '@nestjs/common';
 
 import { DeleteUserService } from './deleteUser.service';
 import { ApiBadRequestResponse, ApiTags } from '@nestjs/swagger';
 
 @ApiTags('Users')
-@Controller('users')
+@Controller('users/:id')
 export class DeleteUserController {
   constructor(private deleteUserService: DeleteUserService) {}
 
@@ -19,7 +21,16 @@ export class DeleteUserController {
   @ApiBadRequestResponse({
     description: 'This will be returned when has validation error',
   })
-  public async create(@Query('id') id: string) {
+  public async create(@Response() res) {
+    const id = res.locals.user;
+
+    if (!id) {
+      throw new HttpException(
+        'User not authenticated!',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
     await this.deleteUserService.delete(id);
+    return res.status(200).send({});
   }
 }
